+++ R/access.R
@@ -99,6 +99,7 @@ is_rmonad <- function(m) {
 # NA_integer_, logical(0), etc) and use of the wrong one can be a source of
 # subtle of reoccuring bugs. So I gather all this into one place.
 .default_value      <- function() void_cache()
+.default_key        <- function() .digest(NULL)
 .default_tag        <- function() ""
 .default_head       <- function() 1L
 .default_code       <- function() character(0)
@@ -110,6 +111,7 @@ is_rmonad <- function(m) {
 .default_mem        <- function() NA_real_
 .default_time       <- function() NA_real_
 .default_meta       <- function() list()
+.default_depth      <- function() 1L
 .default_nest_depth <- function() 1L
 .default_stored     <- function() FALSE
 .default_id         <- function() integer(0)
@@ -162,7 +164,7 @@ has_value <- function(m, ...) {
   sapply(
     .get_many_attributes(m, attribute='value', ...),
     function(x) {
-      (class(x) == "CacheManager") && x@chk()
+      (class(x) == "ValueManager") && x@chk()
     }
   ) %>% unname
 }
@@ -190,41 +192,107 @@ has_summary <- function(m, ...) sapply(get_summary(m, ...), function(x) length(x
 
 # ================================ Tag handling ================================
 
+#' Move head to this id
+#'
+#' @param m rmonad object
+#' @param id integer index
+#' @export
+viewID <- function(m, id){
+  .m_check(m)
+  .check_type(id, type='index', test=function(x) is.numeric(id) && length(id) == 1)
+  m@head <- get_key(m, id)[[1]]
+  m
+}
+
+#' Return a list of Rmonad objects at these positions
+#'
+#' @param m rmonad object
+#' @param ids integer vector index
+#' @export
+viewIDs <- function(m, ids){
+  .m_check(m)
+  .check_type(ids, type='index', test=is.numeric)
+  lapply(ids, viewID, m=m)
+}
+
+.parse_tags <- function(...){
+  tags <- unlist(list(...))
+  tags <- ifelse(tags == "", "/", tags)
+  tags <- unlist(strsplit(tags, '/'))
+  list(tag=tags, str=paste(tags, collapse='/'))
+}
+
 #' Set the head of an Rmonad to a particular tag 
 #'
+#' Will split on '/'
+#'
 #' @param m Rmonad object
-#' @param tag string specifying a single tag for one node in the pipeline 
+#' @param ... one or more tag strings specifying a unique node in the pipeline 
 #' @return Rmonad object with head reset
 #' @export
 #' @examples
 #' library(magrittr)
-#' m <- 256 %v>% sqrt %>% tag('a') %v>% sqrt
-#' esc(view(m, 'a'))
+#' m <- 256 %v>% sqrt %>% tag('a', 'b') %v>% sqrt
+#' esc(view(m, 'a/b'))
 #' funnel(view(m, 'a'), m) %*>% sum
-view <- function(m, tag){
-  tags <- which(sapply(get_tag(m), identical, tag))
+view <- function(m, ...){
+  .m_check(m)
+  x <- .parse_tags(...)
+  tags <- which(sapply(get_tag(m), identical, x$tag))
   if(length(tags) > 1){
-    stop("The given tag, '", tag, "' is ambiguous, maybe use 'views' instead?")
+    msg <- "The given tag, '%s', is ambiguous, maybe use 'views' instead?"
+    stop(sprintf(msg, x$str))
   }
   if(length(tags) == 0){
     msg <- "Tag '%s' not found"
-    stop(sprintf(msg, paste(tag, collapse=", ")))
+    stop(sprintf(msg, x$str))
   }
   m@head <- igraph::vertex_attr(m@graph)$name[tags[1]]
   m
 }
 
+#' Get a list of Rmonad objects matching the given tag
+#'
+#' @param m Rmonad object
+#' @param ... one or more tags
+#' @return list of Rmonad objects
+#' @export
+#' @examples
+#' library(magrittr)
+#' 1 %>>% prod(2) %>% tag('a/b') %>>%
+#'        prod(2) %>% tag('a/c') %>>%
+#'        prod(2) %>% tag('a/c') %>>%
+#'        prod(2) %>% tag('g/a') -> m
+#' views(m, 'a')
+views <- function(m, ...){
+  .m_check(m)
+  x <- .parse_tags(...)
+  node_tags <- get_tag(m)
+  ids <- which(.a_has_prefix_b(node_tags, x$tag))
+  viewIDs(m, ids)
+}
+
 #' Set the tag of an Rmonad object 
 #'
 #' @param m Rmonad object
-#' @param value string specifying a new tag for the given nodes
+#' @param ... one or more tags for the given nodes
 #' @param index character or integer vector, specifying the nodes which will be
 #' assigned the new tag 
 #' @return Rmonad object with new tags
 #' @export
-tag <- function(m, value, index=m@head){
-  .check_type(value, 'character')
-  .set_many_attributes(m, attribute='tag', value=value, index=index)
+#' @examples
+#' library(magrittr)
+#' 1 %>>% prod(2) %>% tag('a/b') %>>% prod(3) %>% get_tag
+#'
+tag <- function(m, ..., index=m@head){
+  x <- .parse_tags(...)
+  if(!is.list(index)){
+    index = list(index)
+  }
+  for(i in index){
+    m <- .set_single_attribute(m, attribute='tag', value=x$tag, index=i)
+  }
+  m
 }
 
 # ============================= Vectorized Getters =============================
@@ -261,6 +329,12 @@ get_prior <- function(m, index=.get_ids(m), tag=NULL) {
 
 #' @rdname rmonad_getters
 #' @export
+get_depth <- function(m, index=.get_ids(m), tag=NULL) {
+  .get_many_attributes(m, index=index, tag=tag, attribute='depth') %>% as.integer
+}
+
+#' @rdname rmonad_getters
+#' @export
 get_nest_depth <- function(m, index=.get_ids(m), tag=NULL) {
   .get_many_attributes(m, index=index, tag=tag, attribute='nest_depth') %>% as.integer
 }
@@ -270,6 +344,12 @@ get_nest_depth <- function(m, index=.get_ids(m), tag=NULL) {
 get_value <- function(m, index=.get_ids(m), tag=NULL, warn=TRUE){
   values <- .get_many_attributes(m, index=index, tag=tag, attribute='value')
   lapply(values, function(v) v@get(warn))
+}
+
+#' @rdname rmonad_getters
+#' @export
+get_key <- function(m, index=.get_ids(m), tag=NULL) {
+  .get_many_attributes(m, index=index, tag=tag, attribute="key")
 }
 
 #' @rdname rmonad_getters
@@ -349,7 +429,7 @@ get_summary <- function(m, index=.get_ids(m), tag=NULL) {
 
 # ============== Public setters ================================================
 # Not all fields SHOULD be settable. For example, I can conceive of no reason
-# why `time` should ever be reset. There are cases where the `CacheManager`
+# why `time` should ever be reset. There are cases where the `ValueManager`
 # object stored in the `value` slot may be changed (for example, to remove a
 # value from cache or change how it is cached), but care must be taken to
 # change only the wrapper, not the pure value. It is possible to set fields
@@ -405,6 +485,14 @@ get_summary <- function(m, index=.get_ids(m), tag=NULL) {
 }
 `.single_value<-` <- function(m, value) {
   .set_single_attribute(m, attribute="value", value=memory_cache(value))
+}
+
+.single_key <- function(m, ...) {
+  .get_single_attribute(m, attribute="key", ...)
+}
+`.single_key<-` <- function(m, value) {
+  .check_type(value, type="md5sum", test=is.character)
+  .set_single_attribute(m, attribute="key", value=value)
 }
 
 .single_raw_value <- function(m, ...){
@@ -496,6 +584,8 @@ get_summary <- function(m, index=.get_ids(m), tag=NULL) {
   .get_single_relative_ids(m, mode="in", type="nest", ...)
 }
 `.single_nest<-` <- function(m, value) {
+  # `value` is the Rmonad that will be nested inside `m`
+  # `value` is the "parent", since its value will be passed to `m`
   if(.single_OK(value)){
     .inherit(
       child         = m,
@@ -517,6 +607,13 @@ get_summary <- function(m, index=.get_ids(m), tag=NULL) {
     .single_raw_value(m) <- void_cache()
     m
   }
+}
+
+.single_depth <- function(m, ...) {
+  .get_single_attribute(m, attribute="depth", ...)
+}
+`.single_depth<-` <- function(m, value) {
+  .set_single_attribute(m, attribute="depth", value=value)
 }
 
 .single_nest_depth <- function(m, ...) {