+++ R/classes.R
@@ -1,12 +1,24 @@
 setClass(
-  "CacheManager",
+  "ValueManager",
   representation(
+    in_memory = "logical",
+    get = "function",
+    del = "function",
+    chk = "function"
+  )
+)
+
+setClass(
+  "Cacher",
+  representation(
+    put = "function",
     get = "function",
     del = "function",
     chk = "function",
-    tag = "function"
+    bld = "function" # build a ValueManager object
   )
 )
+
 
 setOldClass("igraph")
 setClass(
@@ -17,9 +29,9 @@ setClass(
     data = "list"
   )
 )
-Rmonad <- function(){
+Rmonad <- function(node_id){
   m <- new("Rmonad")
-  m <- .new_rmonad_graph(m)
+  m <- .new_rmonad_graph(m, node_id=node_id)
   m@data <- list(RmonadData())
   names(m@data) <- m@head
   m
@@ -28,7 +40,8 @@ Rmonad <- function(){
 setClass(
   "RmonadData",
   representation(
-    value      = "CacheManager",
+    value      = "ValueManager",
+    key        = "character",
     tag        = "character",
     code       = "character",
     error      = "character",
@@ -40,6 +53,7 @@ setClass(
     time       = "numeric",
     meta       = "list",
     summary    = "list",
+    depth      = "integer",
     nest_depth = "integer",
     stored     = "logical",
     options    = "list"
@@ -48,6 +62,7 @@ setClass(
 RmonadData <- function(){
   d <- new("RmonadData")
   d@value      <- .default_value()
+  d@key        <- .default_key()
   d@tag        <- .default_tag()
   d@code       <- .default_code()
   d@error      <- .default_error()