+++ tests/testthat/test-combine.R
@@ -105,10 +105,11 @@ test_that("branches of the same pipeline can be joined", {
   )
 })
 
-test_that("no duplication occurs on pipelines that branch and rejoin", {
+test_that("identical steps are joined", {
   expect_equal(
     {
       256 %v>% sqrt -> m
+      # The `m %v>% sqrt` step is identical
       m %v>% sqrt %v>% sqrt -> a
       m %v>% sqrt -> b
       funnel(a, b) %*>% sum -> ab
@@ -117,7 +118,6 @@ test_that("no duplication occurs on pipelines that branch and rejoin", {
     list(
       256,
       16,
-      NULL, # 4
       4,
       NULL, # 2
       NULL, # funnel
@@ -126,16 +126,38 @@ test_that("no duplication occurs on pipelines that branch and rejoin", {
   )
 })
 
-test_that("Combine with %>>%", {
-  # The history is being lost on functions of type:
-  # m [m a] -> m [a]
-  # The problem
-  # [m a] -> m [a]
-  # works fine
+
+test_that("no duplication occurs on pipelines that branch and rejoin", {
   expect_equal(
-    as_monad(c(1,2,3)) %>% tag('a') %v>%
-      lapply(function(x) { x %>>% sqrt }) %>>% combine %>%
-      get_tag(1),
-    list("a")
+    {
+      256 %v>% sqrt -> m
+      m %v>% prod(10) %v>% prod(100) %>% tag('a') -> a
+      m %v>% prod(2) %>% tag('b') -> b
+      funnel(a, b) %>% tag('ab') %*>% sum -> ab
+      ab %>% get_value(warn=FALSE)
+    },
+    list(
+      256,
+      16,
+      32,
+      160,
+      16000,
+      list(16000, 32),
+      16032
+    )
+  )
+})
+
+test_that("funnel works with view", {
+  expect_equal(
+    2 %>>% prod(2) %>% tag("A") %>>%
+           prod(3) %>% tag("B") %>>%
+           prod(4) %>% tag("C") %>>%
+           prod(5) %>% tag("D") %>%
+             {rmonad::funnel(
+               x = view(., "A"),
+               y = view(., "B")
+             )} %>% esc(T),
+    list(x=4, y=12)
   )
 })