+++ tests/testthat/test-issues.R
@@ -20,7 +20,9 @@ test_that('issue #2: funnel works with %__%', {
 test_that('issue #3: nested errors are localized', {
   expect_equal(
     "a" %>>% {
+      "Level One"
       . %>>% paste("b") %>>% {
+        "Level Two"
         . %>>% paste("c") %>>% stop
       }
     } %>% get_value(warn=FALSE),
@@ -30,7 +32,9 @@ test_that('issue #3: nested errors are localized', {
   expect_equal(
     "yolo" %__%
     "a" %>>% {
+      "Level One"
       . %>>% paste("b") %>>% {
+        "Level Two"
         . %>>% paste("c") %>>% stop
       }
     } %>% get_value(warn=FALSE),