+++ tests/testthat/test-tag.R
@@ -3,6 +3,8 @@ context("tags and views")
 test_that("tag sets tag", {
   # can set tag from as_monad
   expect_equal(as_monad(5, tag='a') %>% get_tag, list("a"))
+  # tag splits on '/'
+  expect_equal(as_monad(5, tag='foo/bar') %>% get_tag, list(c("foo", "bar")))
   # default tag values are empty strings
   expect_equal(
     as_monad(16) %>>% sqrt %>>% sqrt %>% get_tag,
@@ -13,28 +15,35 @@ test_that("tag sets tag", {
     as_monad(16) %>% tag('a') %>>% sqrt %>% tag('b') %>>% sqrt %>% get_tag,
     list("a", "b", "")
   )
+  # tag sets the head node
+  expect_equal(
+    as_monad(16) %>% tag('a', 'foo') %>>% sqrt %>% tag('b') %>>% sqrt %>% get_tag,
+    list(c("a", "foo"), "b", "")
+  )
 })
 
-m <- as_monad(16) %>% tag('a') %>>% sqrt %>% tag('b') %v>% sqrt
+m <- as_monad(16) %>% tag('a') %>>% sqrt %>% tag('foo/bar') %v>% sqrt
 test_that("view gets tagged node", {
-  expect_equal(view(m, 'b') %>% esc, 4)
+  expect_equal(view(m, 'a') %>% esc, 16)
+  expect_equal(view(m, 'foo/bar') %>% esc, 4)
+  expect_equal(view(m, c('foo','bar')) %>% esc, 4)
 })
+
 
 test_that("has_tag works", {
   expect_equal(has_tag(m), c(TRUE, TRUE, FALSE))
 })
 
-m2 <- funnel(view(m, 'a'), view(m, 'b'), 99) %*>% sum
+m2 <- funnel(view(m, 'a'), view(m, 'foo/bar'), 99) %*>% sum
 test_that("view works in funnel", {
   expect_equal(esc(m2), 119)
   expect_equal(get_dependents(m2),    list(5L,   c(3L,5L), c(4L,5L), integer(0), 6L,   integer(0)))
   expect_equal(get_value(m2, warn=F), list(NULL, 16,       4,        2,          NULL, 119       ))
 })
 
-f <- make_recacher(memory_cache)
-m3 <- 'a' %>>% paste('b') %>% f(c('foo', 'bar')) %>>%
-               paste('c') %>% f(c('foo', 'rad')) %>>%
-               paste('d') %>% f('baz') %>>%
+m3 <- 'a' %>>% paste('b') %>% tag(c('foo', 'bar')) %>>%
+               paste('c') %>% tag(c('foo', 'rad')) %>>%
+               paste('d') %>% tag('baz') %>>%
                paste('e')
 test_that("multi level tags work", {
   expect_equal(get_value(m3, tag='foo'), list('foo/bar'='a b', 'foo/rad'='a b c'))
@@ -43,6 +52,7 @@ test_that("multi level tags work", {
 
 test_that("other get_* accessors work", {
   expect_equal(get_OK(m3, tag='foo'), c(T,T))
+  expect_equal(get_OK(m3, tag='foo/bar'), T)
   # I could check the actual return values, but that would probably be
   # overkill. So I just ensure nothing explodes. 
   expect_equal(get_code       (m3, tag='foo') %>% length, 2 )
@@ -63,6 +73,7 @@ test_that("other get_* accessors work", {
   expect_equal(get_value      (m3, tag='foo') %>% length, 2 )
 })
 
+f <- make_recacher(memory_cache)
 m4 <- 256 %>% f(c('a', 'b')) %>>% sqrt %>% f('b') %>>% sqrt %>% f(c('a', 'b'))
 m5 <- 256 %>% f('a') %>>% sqrt %>% f('b') %>>% sqrt %>% f('a')
 test_that("Can search nested tags", {
@@ -70,7 +81,7 @@ test_that("Can search nested tags", {
   expect_equal(names(get_value(m5, tag='a')), c("a/1", "a/3"))
 })
 
-f <- make_recacher(make_local_cacher())
+f <- make_recacher(memory_cache)
 m6 <- 256 %>% f('a') %>>%
   sqrt %>% f('b') %>% funnel(a = ., b=view(., 'a')) %*>%
   sum